+++ packages/agent/CHANGELOG.md
@@ -12,6 +12,30 @@
 
 ### Removed
 
+## [2026.9.12] - 2026-09-12
+
+### Breaking Changes
+
+### Added
+
+### Changed
+
+### Fixed
+
+### Removed
+
+## [2026.9.11] - 2026-09-11
+
+### Breaking Changes
+
+### Added
+
+### Changed
+
+### Fixed
+
+### Removed
+
 ## [2026.9.10-2] - 2026-09-10
 
 ### Breaking Changes